Updates introduction.
It was pointed out that the asynchronity isn't a requirement for the
client if sufficient "bookkeeping" was done. Unfortunately, this
bookkeeping amounts to implementing request tracking with timeouts for
Monto -- something HTTP already has built in.

I therefore believe that this does not fundamentally change the value
proposition of HTTP with error reporting as an alternative to ZeroMQ
with timeouts.